refactor: clearer hierarchy
