+++ arithmetization/src/main/java/net/consensys/linea/rpc/counters/RollupRpcEndpointServicePlugin.java
@@ -13,13 +13,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.services.counters;
+package net.consensys.linea.rpc.counters;
 
 import java.util.Optional;
 
 import com.google.auto.service.AutoService;
 import net.consensys.linea.LineaRequiredPlugin;
-import net.consensys.linea.services.counters.rpc.RollupGenerateCountersV0;
 import net.consensys.linea.zktracer.opcode.OpCodes;
 import org.hyperledger.besu.plugin.BesuContext;
 import org.hyperledger.besu.plugin.BesuPlugin;
