+++ arithmetization/src/main/java/net/consensys/linea/rpc/tracegeneration/RollupRpcEndpointServicePlugin.java
@@ -13,14 +13,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.services.tracegeneration;
+package net.consensys.linea.rpc.tracegeneration;
 
 import java.util.Optional;
 
 import com.google.auto.service.AutoService;
 import lombok.extern.slf4j.Slf4j;
 import net.consensys.linea.LineaRequiredPlugin;
-import net.consensys.linea.services.tracegeneration.rpc.RollupGenerateConflatedTracesToFileV0;
 import net.consensys.linea.zktracer.opcode.OpCodes;
 import org.hyperledger.besu.plugin.BesuContext;
 import org.hyperledger.besu.plugin.BesuPlugin;
